requirements.txt
django
djangorestframework
django-cors-headers
django-environ
stripe
paypal-checkout-serversdk
